refactor(polynomials): isolate factor and subresultant contracts
