+++ tests/math/polynomials/test_multivariate_polynomial.py
@@ -16,8 +16,6 @@ from jacobian.math.polynomials.multivariate._division import (
 from jacobian.math.polynomials.multivariate._gcd import MultivariateGcdRequest
 from jacobian.math.polynomials.multivariate._models import (
     _MAX_MULTIVARIATE_COEFFICIENT_DIGITS,
-    MultivariateSubresultantSequenceRequest,
-    MultivariateSubresultantSequenceResult,
 )
 from jacobian.math.polynomials.multivariate._operations import (
     compute_multivariate_division,
@@ -28,6 +26,10 @@ from jacobian.math.polynomials.multivariate._operations import (
 from jacobian.math.polynomials.multivariate._resultant import (
     MultivariateResultantRequest,
     MultivariateResultantResult,
+)
+from jacobian.math.polynomials.multivariate._subresultants import (
+    MultivariateSubresultantSequenceRequest,
+    MultivariateSubresultantSequenceResult,
 )
 from jacobian.math.polynomials.values import RationalPolynomial
